fix: abstract code and improve documentation
