+++ src/Progress/index.tsx
@@ -4,6 +4,7 @@ import ProgressLine from './ProgressLine';
 
 // export types
 export type { ProgressLineProps as ProgressProps } from './ProgressLine';
+export type { ProgressSection } from './types';
 
 // export components
 export { Progress, ProgressCircle, ProgressLine };
