+++ src/ProgressCircle/ProgressCircle.tsx
@@ -1,9 +1,11 @@
 import React, { useMemo } from 'react';
 import Box, { BoxProps } from '@/internals/Box';
+import ProgressInfo from '../Progress/ProgressInfo';
+import ProgressCircleSections from './ProgressCircleSections';
+import useProgressCirclePath from './hooks/useProgressCirclePath';
 import { forwardRef } from '@/internals/utils';
 import { useStyles, useCustom } from '@/internals/hooks';
-import ProgressInfo from '../Progress/ProgressInfo';
-import { ProgressSection } from '../Progress/ProgressLine';
+import type { ProgressSection } from '../Progress';
 
 export interface ProgressCircleProps extends BoxProps {
   /** Circular progress bar degree */
@@ -36,11 +38,11 @@ export interface ProgressCircleProps extends BoxProps {
   /** Tail width */
   trailWidth?: number;
 
+  /** Multiple sections with different colors */
+  sections?: Pick<ProgressSection, 'percent' | 'color'>[];
+
   /** Custom render function for info content */
   renderInfo?: (percent: number, status?: 'success' | 'fail' | 'active') => React.ReactNode;
-
-  /** Multiple sections with different colors */
-  sections?: ProgressSection[];
 }
 
 /**
@@ -78,58 +80,14 @@ const ProgressCircle = forwardRef<'div', ProgressCircleProps>((props, ref) => {
     );
   }, [percent, sections]);
 
-  const { pathString, trailPathStyle, strokePathStyle } = useMemo(() => {
-    const radius = 50 - strokeWidth / 2;
-
-    let x1 = 0;
-    let y1 = -radius;
-    let x2 = 0;
-    let y2 = -2 * radius;
-
-    switch (gapPosition) {
-      case 'left':
-        x1 = -radius;
-        y1 = 0;
-        x2 = 2 * radius;
-        y2 = 0;
-        break;
-      case 'right':
-        x1 = radius;
-        y1 = 0;
-        x2 = -2 * radius;
-        y2 = 0;
-        break;
-      case 'bottom':
-        y1 = radius;
-        y2 = 2 * radius;
-        break;
-      default:
-    }
-
-    const pathString = `M 50,50 m ${x1},${y1} a ${radius},${radius} 0 1 1 ${x2},${-y2} a ${radius},${radius} 0 1 1 ${-x2},${y2}`;
-
-    const len = Math.PI * 2 * radius;
-    // Convert gapDegree from degrees to a proportion of the circumference
-    const gapLength = (gapDegree / 360) * len;
-
-    const trailPathStyle = {
-      stroke: trailColor,
-      strokeDasharray: `${len - gapLength}px ${len}px`,
-      strokeDashoffset: `-${gapLength / 2}px`
-    };
-
-    const strokePathStyle = {
-      stroke: strokeColor,
-      strokeDasharray: `${(totalPercent / 100) * (len - gapLength)}px ${len}px`,
-      strokeDashoffset: `-${gapLength / 2}px`
-    };
-
-    return {
-      pathString,
-      trailPathStyle,
-      strokePathStyle
-    };
-  }, [gapDegree, gapPosition, totalPercent, strokeColor, strokeWidth, trailColor]);
+  const { pathString, trailPathStyle, strokePathStyle } = useProgressCirclePath({
+    gapDegree,
+    gapPosition,
+    totalPercent,
+    strokeColor,
+    strokeWidth,
+    trailColor
+  });
 
   const { prefix, merge, withPrefix } = useStyles(classPrefix);
   const classes = merge(className, withPrefix({ [`${status || ''}`]: !!status }));
@@ -163,40 +121,15 @@ const ProgressCircle = forwardRef<'div', ProgressCircleProps>((props, ref) => {
           style={trailPathStyle}
         />
         {sections ? (
-          // Render multiple sections
-          <>
-            {(() => {
-              let startPercent = 0;
-              return sections.map((section, index) => {
-                const sectionLen = Math.PI * 2 * (50 - strokeWidth / 2);
-                const gapLength = (gapDegree / 360) * sectionLen;
-                const sectionPercent = section.percent;
-                const endPercent = startPercent + sectionPercent;
-
-                // Calculate the stroke dash array and offset for this section
-                const sectionStyle = {
-                  stroke: section.color,
-                  strokeDasharray: `${(sectionPercent / 100) * (sectionLen - gapLength)}px ${sectionLen}px`,
-                  strokeDashoffset: `-${gapLength / 2 + (startPercent / 100) * (sectionLen - gapLength)}px`
-                };
-
-                const sectionPath = (
-                  <path
-                    key={index}
-                    d={pathString}
-                    strokeLinecap={strokeLinecap}
-                    className={prefix('stroke')}
-                    strokeWidth={totalPercent === 0 ? 0 : strokeWidth}
-                    fillOpacity="0"
-                    style={sectionStyle}
-                  />
-                );
-
-                startPercent = endPercent;
-                return sectionPath;
-              });
-            })()}
-          </>
+          <ProgressCircleSections
+            classPrefix={classPrefix}
+            sections={sections}
+            pathString={pathString}
+            strokeLinecap={strokeLinecap}
+            strokeWidth={strokeWidth}
+            gapDegree={gapDegree}
+            totalPercent={totalPercent}
+          />
         ) : (
           // Render single stroke
           <path
